+++ tests/testthat.R
@@ -1,9 +1,7 @@
 library(testthat)
 library(rbmi)
 
-if (Sys.getenv("NOT_CRAN") != "true") {
-    Sys.setenv(RBMI_ENABLE_CACHE = "false")
-    options(rbmi.enable_cache = FALSE)
-}
+# Disable cache by default during unit tests
+options("rbmi.enable_cache" = FALSE)
 
 test_check("rbmi")